+++ CHANGELOG.md
@@ -17,7 +17,7 @@ unreleased(TODO)
 * a little bit refactoring
 - * add html5 inputs
 * added message for should\_not
-- * raise exception when wrong parametres specified(:count and :minimum simultaneously)
+* raise exception when wrong parametres specified(:count and :minimum (or :maximum) simultaneously)
 * support all versions of nokogiri since 1.4.4
 
 0.1.6 (nokogiri update)